% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compatibility.R
\name{as.IdMap}
\alias{as.IdMap}
\title{Perform the transformation of the UniquePairs object into the IdMap object}
\usage{
as.IdMap(uniquePairs, keepOrder = TRUE, verbose = FALSE)
}
\arguments{
\item{uniquePairs}{the UniquePairs object to be converted into IdMap}

\item{keepOrder}{logical indicating if the original order of primary IDs should be lept. Default is TRUE}

\item{verbose}{if TRUE enables diagnostic messages}
}
\description{
Perform the transformation of the UniquePairs object into the IdMap object
}
\author{
Roger Day, Alex Lisovich
}
\keyword{internal}

