%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationNetAffx.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{filesExist.AnnotationNetAffx}
\alias{filesExist.AnnotationNetAffx}
\alias{AnnotationNetAffx.filesExist}
\alias{filesExist.AnnotationNetAffx}
\alias{filesExist,AnnotationNetAffx-method}


 \title{Check if the AnnotationNetAffx contains the reference to an existing data file or directory}
 \description{Check if the AnnotationNetAffx contains the reference to an existing data file or directory.}

 \usage{\method{filesExist}{AnnotationNetAffx}(this, ...)}

 \arguments{
 \item{...}{Not used}
 }

 \value{\code{\link[base:logical]{TRUE}} if file(s) or directory exists, FALSE otherwise.}

 \seealso{For more information see \code{\link{AnnotationNetAffx}}.}

 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
