\name{Generic R.00 object methods}
\alias{Generic R.00 object methods}
\title{Generic R.00 object methods}

\alias{addServices}
\alias{createSubmission}
\alias{filesExist}
\alias{getArrayType}
\alias{getArrayTypes}
\alias{getColumns}
\alias{getColumns1}
\alias{getColumns2}
\alias{getCredentials}
\alias{getDataFrame}
\alias{getDataFrameList}
\alias{getDefaultServices}
\alias{getFolderName}
\alias{getIdMap}
\alias{getIdMapList}
\alias{getProbesetList}
\alias{getRoot}
\alias{getServiceRoot}
\alias{getServices}
\alias{init}
\alias{readDF}
\alias{setCredentials}
\alias{setOptions}
\alias{setServices}

\description{

This page contains generic functions that call UseMethod() on S3 classes defined within the R.oo framework.
For detailed description, please refer to the particular document for the corresponding method coupled to a specific class.

}

\keyword{documentation}
\keyword{internal}
