%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationAffx.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColumns.AnnotationAffx}
\alias{getColumns.AnnotationAffx}
\alias{AnnotationAffx.getColumns}
\alias{getColumns.AnnotationAffx}
\alias{getColumns,AnnotationAffx-method}


\title{Extract ID match pairs from the data frame obtained from the Affymetrix online data repository}

\description{
 See \code{\link[IdMappingRetrieval:getColumns.Annotation]{getColumns}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationAffx}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
