%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationDavid.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColumns.AnnotationDavid}
\alias{getColumns.AnnotationDavid}
\alias{AnnotationDavid.getColumns}
\alias{getColumns.AnnotationDavid}
\alias{getColumns,AnnotationDavid-method}


\title{Extract ID match pairs from the data frame obtained from the DAVID online query system}

\description{
 See \code{\link[IdMappingRetrieval:getColumns.Annotation]{getColumns}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationDavid}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
