%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationDavidCsv.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColumns.AnnotationDavidCsv}
\alias{getColumns.AnnotationDavidCsv}
\alias{AnnotationDavidCsv.getColumns}
\alias{getColumns.AnnotationDavidCsv}
\alias{getColumns,AnnotationDavidCsv-method}


\title{Extract ID match pairs from the data frame obtained from the DAVID backend conversion results file}

\description{
 See \code{\link[IdMappingRetrieval:getColumns.Annotation]{getColumns}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationDavidCsv}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
