%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationEnsemblCsv.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColumns.AnnotationEnsemblCsv}
\alias{getColumns.AnnotationEnsemblCsv}
\alias{AnnotationEnsemblCsv.getColumns}
\alias{getColumns.AnnotationEnsemblCsv}
\alias{getColumns,AnnotationEnsemblCsv-method}


\title{Extract ID match pairs from the data frame obtained from the Ensembl online query system}

\description{
 See \code{\link[IdMappingRetrieval:getColumns.Annotation]{getColumns}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationEnsemblCsv}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
