%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationEnvision.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColumns.AnnotationEnvision}
\alias{getColumns.AnnotationEnvision}
\alias{AnnotationEnvision.getColumns}
\alias{getColumns.AnnotationEnvision}
\alias{getColumns,AnnotationEnvision-method}


\title{Extract ID match pairs from the data frame obtained from the Envision online query system}

\description{
 See \code{\link[IdMappingRetrieval:getColumns.Annotation]{getColumns}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationEnvision}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
