%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFolderName.Annotation}
\alias{getFolderName.Annotation}
\alias{Annotation.getFolderName}
\alias{getFolderName.Annotation}
\alias{getFolderName,Annotation-method}


\title{Get caching folder name for a given Annotation object}

\description{
 Get caching folder name for a given Annotation object. To retrieve a full path for a given service caching folder,
 see \code{\link[IdMappingRetrieval:getServiceRoot.Annotation]{getServiceRoot}}.
}

\usage{\method{getFolderName}{Annotation}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{A \code{\link[base]{character}} string containing caching folder name.}

\seealso{For more information see \code{\link{Annotation}}.}

\author{
 Alex Lisovich, Roger Day
}
\keyword{internal}
\keyword{methods}
