% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotationDataRoot.R
\docType{data}
\name{IdMappingAnnotationData.root}
\alias{IdMappingAnnotationData.root}
\title{Annotation data root directory}
\description{
Annotation data root directory
}
\author{
Alex Lisovich, Roger Day
}
\keyword{data}
\keyword{internal}

