% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IdMappingRetrieval-package.R
\docType{package}
\name{IdMappingRetrieval-package}
\alias{IdMappingRetrieval}
\alias{IdMappingRetrieval-package}
\title{Id Mapping Retrieval.}
\description{
\tabular{ll}{
Package: \tab IdMappingRetrieval\cr
Type: \tab Package\cr
Version: \tab 0.99.0 \cr
Date: \tab 2011-06-26 \cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alex Lisovich, Roger Day
}
\keyword{package}

