%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationAffx.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationAffx}
\alias{readDF.AnnotationAffx}
\alias{AnnotationAffx.readDF}
\alias{readDF.AnnotationAffx}
\alias{readDF,AnnotationAffx-method}


\title{Retrieve the annotation data frame from the service data repository}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationAffx}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
