%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationDavid.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationDavid}
\alias{readDF.AnnotationDavid}
\alias{AnnotationDavid.readDF}
\alias{readDF.AnnotationDavid}
\alias{readDF,AnnotationDavid-method}


\title{Retrieve the annotation data frame from the DAVID online query system}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationDavid}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
