%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationDavidCsv.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationDavidCsv}
\alias{readDF.AnnotationDavidCsv}
\alias{AnnotationDavidCsv.readDF}
\alias{readDF.AnnotationDavidCsv}
\alias{readDF,AnnotationDavidCsv-method}


\title{Retrieve the annotation data frame from the DAVID backend conversion results file}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationDavidCsv}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
