%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationEnsembl.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationEnsembl}
\alias{readDF.AnnotationEnsembl}
\alias{AnnotationEnsembl.readDF}
\alias{readDF.AnnotationEnsembl}
\alias{readDF,AnnotationEnsembl-method}


\title{Retrieve the annotation data frame from the Ensembl online query system}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationEnsembl}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
