%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationEnsemblCsv.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationEnsemblCsv}
\alias{readDF.AnnotationEnsemblCsv}
\alias{AnnotationEnsemblCsv.readDF}
\alias{readDF.AnnotationEnsemblCsv}
\alias{readDF,AnnotationEnsemblCsv-method}


\title{Retrieve the annotation data frame from the Ensembl csv query results file}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationEnsemblCsv}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
