%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationEnvision.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationEnvision}
\alias{readDF.AnnotationEnvision}
\alias{AnnotationEnvision.readDF}
\alias{readDF.AnnotationEnvision}
\alias{readDF,AnnotationEnvision-method}


\title{Retrieve the annotation data frame using Envision online query system}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationEnvision}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
