%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  annotationNetAffx.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDF.AnnotationNetAffx}
\alias{readDF.AnnotationNetAffx}
\alias{AnnotationNetAffx.readDF}
\alias{readDF.AnnotationNetAffx}
\alias{readDF,AnnotationNetAffx-method}


\title{Retrieve the annotation data frame from the service data repository}

\description{
 See \code{\link[IdMappingRetrieval:readDF.Annotation]{readDF}} in \code{\link{Annotation}}.
}

\seealso{For more information see \code{\link{AnnotationNetAffx}}.}

\author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
