% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compatibility.R
\name{swapKeys}
\alias{swapKeys}
\title{Swap the primary and secondary key columns}
\usage{
swapKeys(idMap, verbose = FALSE)
}
\arguments{
\item{idMap}{the IdMap object or a list of IdMap objects which keys to be swapped}

\item{verbose}{if TRUE enables diagnostic messages}
}
\value{
IdMap object or list of IdMap objects with swapped keys
}
\description{
Swap the primary and secondary key columns
}
\author{
Roger Day, Alex Lisovich
}
\keyword{internal}

