\name{as.IdMap}
\alias{as.IdMap}
\title{Perform the transformation of the UniquePairs object into the IdMap object...}
\usage{as.IdMap(uniquePairs, keepOrder=TRUE, verbose=FALSE)}
\description{Perform the transformation of the UniquePairs object into the IdMap object}
\keyword{internal}
\author{Roger Day, Alex Lisovich}
\arguments{\item{uniquePairs}{the UniquePairs object to be converted into IdMap}
\item{keepOrder}{logical indicating if the original order of primary IDs should be lept. Default is TRUE}
\item{verbose}{if TRUE enables diagnostic messages}}
