\name{IdMappingRetrieval-package}
\title{Id Mapping Retrieval.}
\description{Id Mapping Retrieval.}
\details{\tabular{ll}{
Package: \tab IdMappingRetrieval\cr
Type: \tab Package\cr
Version: \tab 0.99.0 \cr
Date: \tab 2011-06-26 \cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}}
\alias{IdMappingRetrieval}
\docType{package}
\author{Alex Lisovich, Roger Day}
\keyword{package}
