\name{swapKeys}
\alias{swapKeys}
\title{Swap the primary and secondary key columns...}
\usage{swapKeys(idMap, verbose=FALSE)}
\description{Swap the primary and secondary key columns}
\value{IdMap object or list of IdMap objects with swapped keys}
\keyword{internal}
\author{Roger Day, Alex Lisovich}
\arguments{\item{idMap}{the IdMap object or a list of IdMap objects which keys to be swapped}
\item{verbose}{if TRUE enables diagnostic messages}}
